+++ angular_compiler/CHANGELOG.md
@@ -1,5 +1,7 @@
 *   Catches an (invalid) `null` token of a provider and throws a better error.
 
+*   Catches an (invalid) `null` value of the function for `FactoryProvider`.
+
 ## 0.4.0
 
 ### New Features